tts_worker/requirements.txt
fastapi==0.115.6
uvicorn==0.30.0
pydantic==2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
boto3>=1.34.0
numpy==1.26.4
soundfile==0.12.1
